+++ assets/src/components/activities/short_answer/ShortAnswerAuthoring.tsx
@@ -31,6 +31,7 @@ import { ShortAnswerActions } from './actions';
 import { ShortAnswerModelSchema } from './schema';
 import { VariableEditorOrNot } from '../common/variables/VariableEditorOrNot';
 import { VariableActions } from '../common/variables/variableActions';
+import { Explanation } from '../common/explanation/ExplanationAuthoring';
 
 const store = configureStore();
 
@@ -118,6 +119,9 @@ const ShortAnswer = () => {
 
         <TabbedNavigation.Tab label="Hints">
           <Hints partId={model.authoring.parts[0].id} />
+        </TabbedNavigation.Tab>
+        <TabbedNavigation.Tab label="Explanation">
+          <Explanation partId={DEFAULT_PART_ID} />
         </TabbedNavigation.Tab>
         <TabbedNavigation.Tab label="Dynamic Variables">
           <VariableEditorOrNot